fix: 기본이미지 S3에서 삭제 안되도록 수정

+++ src/main/java/org/example/ctrlu/domain/user/application/UserService.java
@@ -55,7 +55,10 @@ public class UserService {
 			profileImageKey = defaultImageKey;
 		}
 
-		awsS3Service.deleteImage(user.getProfileImageKey());
+		if (!user.getProfileImageKey().equals(defaultImageKey)) {
+			awsS3Service.deleteImage(user.getProfileImageKey());
+		}
+
 		user.updateProfile(request.nickname(), profileImageKey);
 	}
